Folks google-closure code for sourcemap support.
This commit takes the sourcemap package from google-closure project to add better support for sourcemap in the future to WUIC. The code has been modified to be compliant with Java 6 and to not depend on guava. New dependencies to "gson" and "protobuf" projects have been added to satisfy the other dependencies of the imported classes.
According to the Apache 2 licence, explicit quotation of the project has been added to the code and the documentation of the project.
Those modifications are mainly done to fix #167. Currently WUIC does not rely on the imported package and integration of the closure classes has to be done in the future.