Do not perform replacement on the fly.
Each LineInspector should perform its replacement on the fly because they can break the nuts positions inside a combined one.
Those information provided by the CompositeInputStream are not broken if the replacement is performed from the end to the beginning of the content.
To do this, ReplacementInfo class has been modified to be ordered in that way. Replacement is performed only when all ReplacementInfo have been built.
Fixes #210.