Identifies nut position inside a combined nut by counting characters, not bytes.
The CompositeNut tracks position of each nut inside the aggregated content. However, positions where computed by counting byte ranges of each nut inside the aggregated content, while it's expected for text stream  to count characters. In fact, issues can occurs when some characters are encoded on several bytes. Now, the CompositeNut checks if the content is a text stream and in that case it counts characters. Otherwise it counts byte. Fixes #209.