Introduces new implementation for CSS/JS script inspections.
This commit introduces the ScriptLineInspector that fixes #193. The approach of this inspector is to analyze a character stream to detect comment and string literals and submit chunk of data to the legacy regex-based inspectors. In this way, a regex-based inspector will just receive the right data to analyze, and won't for instance match a string inside a literal while it should be the case only in comments.  The commit also complements the commit 718bd112a3b20318f98fcb9c9b1cbf15d51cd22a by using the default platform charset in core engines.