Configures component with method instead of constructors.
@ConfigConstructor annotation has been renamed to @Config and now applies to both methods and constructors in a managed class.
Multiple methods can be annotated with @Config, they will be invoked just after the instantiation of the constructor.
All components in the project have been refactored to be initialized with methods in order to expose a default constructor.
This will allow to delegate instantiation to a ServiceLoader.
Fixes https://github.com/wuic/wuic/issues/221