Removes annotation-detector and use ServiceLoader instead.
Since annotation-detector was not easy to run everywhere (see #216), it has been decided to fallback to the standard mechanism of the JDK: ServiceLoader.
This require a little bit more configuration (basically declare the implementation in META-INF/services), but this cost is justified by the better guarantee given by this implementation in terms of portability. This change technically fixes #216.
Fixes #184.