Simplifies the PropertyResolver interface.
The getKeys() method can be tricky to implement in some case. That's why has been removed from the PropertyResolver interface.
This change is balanced by the enhancement of ObjectBuilder interface providing the list of supported property keys.
This method allows the TaggedSettings to know what property should be retrieved from the PropertyResolver without the need to try to apply each available property.