Transformer is now allowed to not copy stream.
The Transformer interface has been modified by changing the transform signature from void to a returned boolean value.
When the returned value is true, the Pipe considers that something has been written to the given output stream and should be used as input stream to the next transformer.
If the value is false, the Pipe just ignores the output stream and uses the original input stream for the next transformer.
This allows any transformer to not copy the stream if it has nothing to do, improving performances.