Removes writes capabilities from NutDao.
WUIC API was designed to let a workflow write the result of a transformation somewhere thanks to a NutDao.
This feature is not supporting an approach matching basic architecture principles because we almost never allow the app server to reach the reverse proxy.
Moreover, save capability was implemented nowhere and the API was never concretely used.
This commit removes this code not really addressing a need and taking maintenance effort.

Closes #36