Reports statistics via JMX.
This commit published via JMX the last (1000 by default) statistics of workflow executions and heap resolutions.
Workflow executions are measuring the the time spent by the engine to parse the request and the time to transform.
Since parsing process can lead to the measurement of other parsing and even transformation sub process, a special TimerTree object is used to ignore the time spent by those sub process when reporting the parent statistics.
Heap resolution are the time actually spent by the checkFiles() method of the NutsHeap class that call the DAO to resolve the nuts, which is the time consuming operations interesting to measure.

Fixes #27