Adds supports JMX operations.
The WuicFacade now configures a PropertyResolver invoked before all the others that checks if the property is defined in an internal map that can be populated via JMX.
When the JMX operation is invoked, the ContextBuilder is completely refreshed and expires the built context. In order to take into consideration any property, the cached components are also free.

Fixes #60