Refactor version number callback registration mechanism.
A generic version number callback based on the engine's builder properties is now registered when the engine is built.
This allows to simplify NodeEngine's subclasses that don't have to specify the callback implementation and is more robust because all the properties are taken into consideration, making sure that none of them are missed.